Implement chat last message (not complete yet)
Message timestamp and status is not implemented yet.
It is specified that the last message should be shown and how it should look like, but not how to convert a message object to text.
So it will not work. Conversion of some message types is implemented but not committed yet.