Refactor chat photos and fix a serious performance bug
RAM and CPU consumption increases exponentially with chats count, to the point that several apps crashed in my potato laptop.

The culprit:
Every time an `updateChatPosition` comes (when loading chats and when a chat moves), the chat list is re-rendered. (normal here)
However, on each re-render, ALL chat photos were reread from local storage, which was unnecessary as they are already stored in memory.